refactor(search-view): extract useSearchUi hook
